# projects/dev-plm/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dev-plm 在线协作 API · 设计契约",
    "description": "仅设计，未实现后端、数据库或协作客户端。项目事实源于文件；在线数据库提供派生读模型，讨论与认领可为数据库真源并须导出。离线站不依赖此 API。",
    "version": "0.3.0"
  },
  "servers": [
    {
      "url": "https://dev.dev-plm.example.invalid",
      "description": "开发设计地址，未部署",
      "x-environment": "development"
    },
    {
      "url": "https://test.dev-plm.example.invalid",
      "description": "测试设计地址，未部署",
      "x-environment": "testing"
    },
    {
      "url": "https://dev-plm.example.invalid",
      "description": "生产设计地址，未部署",
      "x-environment": "production"
    }
  ],
  "paths": {
    "/v1/auth/token": {
      "post": {
        "tags": [
          "身份认证"
        ],
        "summary": "取得在线协作访问令牌",
        "description": "自定义对接设计，最终身份提供方待实现阶段确认；不声称是 OAuth。租户由认证身份与授权映射确定，不从任意请求正文信任。",
        "operationId": "issueAccessToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              },
              "examples": {
                "design": {
                  "summary": "设计示例；无在线服务，所有认证占位值无效",
                  "value": {
                    "identifier": "DESIGN-USER",
                    "credential": "DESIGN_ONLY_NOT_A_CREDENTIAL"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "设计响应；不代表服务已运行。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                },
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "access_token": "DESIGN_ONLY_NOT_A_VALID_TOKEN",
                      "token_type": "Bearer",
                      "expires_in": 900
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "认证材料无效，不能换取令牌。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "INVALID_CREDENTIALS",
                      "message": "认证材料无效，不能换取令牌。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "输入不符合契约。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "VALIDATION_ERROR",
                      "message": "输入不符合契约。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "明确错误码与请求回执",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-requirements": [
          "DEV-REQ-011",
          "DEV-REQ-012"
        ],
        "x-screens": [],
        "x-fields": [
          "DEV-FLD-TENANT-ID"
        ]
      }
    },
    "/v1/projects": {
      "get": {
        "tags": [
          "项目事实"
        ],
        "summary": "列出可访问项目",
        "description": "设计契约；本轮无服务实现。",
        "operationId": "listProjects",
        "responses": {
          "200": {
            "description": "设计响应；不代表服务已运行。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Projects"
                },
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "items": [
                        {
                          "id": "dev-plm",
                          "name": "dev-plm",
                          "type": "software",
                          "example": false,
                          "stage": "第3轮进行中",
                          "source_path": "projects/dev-plm/project.yaml"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "没有有效访问令牌。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "AUTH_REQUIRED",
                      "message": "没有有效访问令牌。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "对象不存在或对当前身份不可见。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "NOT_FOUND",
                      "message": "对象不存在或对当前身份不可见。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "明确错误码与请求回执",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-requirements": [
          "DEV-REQ-006",
          "DEV-REQ-008",
          "DEV-REQ-009",
          "DEV-REQ-012"
        ],
        "x-screens": [
          "DEV-UI-PROJECTS"
        ],
        "x-fields": [
          "DEV-FLD-PROJECT-ID",
          "DEV-FLD-PROJECT-TYPE",
          "DEV-FLD-TENANT-ID",
          "DEV-FLD-SOURCE-PATH"
        ]
      }
    },
    "/v1/projects/{project_id}/requirements": {
      "get": {
        "tags": [
          "项目事实"
        ],
        "summary": "读取项目需求原文与状态",
        "description": "设计契约；本轮无服务实现。",
        "operationId": "listRequirements",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Project Id"
            },
            "examples": {
              "design": {
                "value": "dev-plm"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设计响应；不代表服务已运行。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requirements"
                },
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "items": [
                        {
                          "id": "DEV-REQ-010",
                          "title": "从用户原话看到真实改动",
                          "original": "每次写下新提示词，Agent 对我的项目做了哪些调整、具体在哪个环节变了，要能看见",
                          "status": "本轮进行中",
                          "source_path": "projects/dev-plm/requirements/DEV-REQ-010.md"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "没有有效访问令牌。",
            "content": {
              "application/json": {
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "AUTH_REQUIRED",
                      "message": "没有有效访问令牌。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "对象不存在或对当前身份不可见。",
            "content": {
              "application/json": {
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "NOT_FOUND",
                      "message": "对象不存在或对当前身份不可见。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "明确错误码与请求回执",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "明确错误码与请求回执",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-requirements": [
          "DEV-REQ-003",
          "DEV-REQ-008",
          "DEV-REQ-012"
        ],
        "x-screens": [
          "DEV-UI-REQUIREMENTS"
        ],
        "x-fields": [
          "DEV-FLD-PROJECT-ID",
          "DEV-FLD-REQUIREMENT-ID",
          "DEV-FLD-ORIGINAL",
          "DEV-FLD-STATUS",
          "DEV-FLD-SOURCE-PATH"
        ]
      }
    },
    "/v1/requirements/{requirement_id}/changes": {
      "get": {
        "tags": [
          "项目事实"
        ],
        "summary": "读取需求关联变更",
        "description": "设计契约；本轮无服务实现。",
        "operationId": "listRequirementChanges",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Requirement Id"
            },
            "examples": {
              "design": {
                "value": "DEV-REQ-008"
              }
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设计响应；不代表服务已运行。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Changes"
                },
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "items": [
                        {
                          "id": "CHG-002",
                          "title": "第2轮：文件驱动的多项目工作台",
                          "requirements": [
                            "DEV-REQ-005",
                            "DEV-REQ-006",
                            "DEV-REQ-008"
                          ],
                          "commits": [
                            "d126b25ef7e4813801d8dfada7ec6213e766b4c4"
                          ],
                          "narrative": "变更单作者自述，详见源文件；不替代下方 Git 事实。",
                          "source_path": "projects/dev-plm/changes/CHG-002.md"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "没有有效访问令牌。",
            "content": {
              "application/json": {
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "AUTH_REQUIRED",
                      "message": "没有有效访问令牌。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "对象不存在或对当前身份不可见。",
            "content": {
              "application/json": {
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "NOT_FOUND",
                      "message": "对象不存在或对当前身份不可见。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "明确错误码与请求回执",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "明确错误码与请求回执",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-requirements": [
          "DEV-REQ-003",
          "DEV-REQ-010",
          "DEV-REQ-012"
        ],
        "x-screens": [
          "DEV-UI-REQUIREMENTS",
          "DEV-UI-CHANGES"
        ],
        "x-fields": [
          "DEV-FLD-REQUIREMENT-ID",
          "DEV-FLD-CHANGE-ID",
          "DEV-FLD-COMMIT-SHA"
        ]
      }
    },
...
    "/v1/sync-jobs/{job_id}": {
      "get": {
        "tags": [
          "文件同步"
        ],
        "summary": "读取同步状态",
        "description": "设计契约；本轮无服务实现。",
        "operationId": "getProjection",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Job Id"
            },
            "examples": {
              "design": {
                "value": "DESIGN-SYNC-001"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设计响应；不代表服务已运行。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncJob"
                },
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "id": "DESIGN-SYNC-001",
                      "state": "queued",
                      "source_snapshot": "DESIGN-SNAPSHOT-NOT-A-REAL-HASH",
                      "message": "设计示例：已接收同步请求，尚未完成校验与发布。"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "没有有效访问令牌。",
            "content": {
              "application/json": {
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "AUTH_REQUIRED",
                      "message": "没有有效访问令牌。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "对象不存在或对当前身份不可见。",
            "content": {
              "application/json": {
                "examples": {
                  "design": {
                    "summary": "设计示例；无在线服务，所有认证占位值无效",
                    "value": {
                      "code": "NOT_FOUND",
                      "message": "对象不存在或对当前身份不可见。",
                      "request_id": "DESIGN-REQUEST-001"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "明确错误码与请求回执",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "明确错误码与请求回执",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-requirements": [
          "DEV-REQ-008",
          "DEV-REQ-012"
        ],
        "x-screens": [],
        "x-fields": [
          "DEV-FLD-SOURCE-HASH",
          "DEV-FLD-JOB-STATE"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessToken": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "const": "Bearer",
            "title": "Token Type"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expires In"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "title": "AccessToken"
      },
...
      "Changes": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Change"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Changes"
      },
...
      "Projects": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Projects"
      },
...
      "Requirements": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Requirements"
      },
...
      "TokenRequest": {
        "properties": {
          "identifier": {
            "type": "string",
            "minLength": 1,
            "title": "Identifier"
          },
          "credential": {
            "type": "string",
            "minLength": 1,
            "title": "Credential",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identifier",
          "credential"
        ],
        "title": "TokenRequest"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "短期访问令牌；租户与角色从已验证身份获得。",
        "scheme": "bearer"
      }
    }
  },
...
}